@@ -18,21 +18,19 @@
     "build": "ng build --prod --sourceMap=false",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e",
-    "postinstall": "npm run build"
+    "e2e": "ng e2e"
   },
   "dependencies": {
-    "@angular/animations": "^7.2.0",
+    "@angular/animations": "^8.2.14",
     "@angular/cdk": "^7.2.1",
-    "@angular/common": "^7.2.0",
-    "@angular/compiler": "^7.2.0",
-    "@angular/core": "^7.2.0",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
     "@angular/flex-layout": "^7.0.0-beta.22",
-    "@angular/forms": "^7.2.0",
-    "@angular/http": "^7.2.0",
-    "@angular/platform-browser": "^7.2.0",
-    "@angular/platform-browser-dynamic": "^7.2.0",
-    "@angular/router": "^7.2.0",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@ng-select/ng-select": "^2.13.3",
     "@ngx-translate/core": "^11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
@@ -50,23 +48,23 @@
     "ngx-device-detector": "^1.3.5",
     "ngx-toastr": "10.1.0",
     "popper.js": "^1.14.6",
-    "rxjs": "^6.3.3",
+    "rxjs": "^6.6.6",
     "rxjs-compat": "^6.3.3",
     "tslib": "^1.9.3",
-    "zone.js": "^0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.21",
-    "@angular/cli": "^7.1.4",
-    "@angular/compiler-cli": "^7.2.0",
-    "@angular/language-service": "^7.2.0",
+    "@angular-devkit/build-angular": "~0.803.29",
+    "@angular/cli": "^8.3.29",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@types/jasmine": "~3.3.5",
     "@types/jasminewd2": "~2.0.6",
     "@types/jsonwebtoken": "^8.3.5",
     "@types/jwt-decode": "^2.2.1",
     "@types/lodash": "^4.14.119",
     "@types/node": "~10.12.18",
-    "codelyzer": "^4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~3.1.4",
@@ -78,6 +76,6 @@
     "protractor": "~5.4.2",
     "ts-node": "~7.0.1",
     "tslint": "~5.12.0",
-    "typescript": "~3.2.2"
+    "typescript": "~3.5.3"
   }
 }
